enter and tab

+++ src/editor/view.rs
@@ -32,7 +32,8 @@ impl View {
             EditorCommand::Quit => {}
             EditorCommand::Insert(character) => self.insert_char(character),
             EditorCommand::Delete => self.delete(),
-            EditorCommand::Backspace => self.backspace(),
+            EditorCommand::Backspace => self.delete_backward(),
+            EditorCommand::Enter => self.insert_newline(),
         }
     }
     pub fn load(&mut self, file_name: &str) {
@@ -48,9 +49,16 @@ impl View {
         self.needs_redraw = true;
     }
 
-    fn backspace(&mut self) {
-        self.move_left();
-        self.delete();
+    fn insert_newline(&mut self) {
+        self.buffer.insert_newline(self.text_location);
+        self.move_text_location(&Direction::Right);
+        self.needs_redraw = true;
+    }
+    fn delete_backward(&mut self) {
+        if self.text_location.line_index != 0 || self.text_location.grapheme_index != 0 {
+            self.move_text_location(&Direction::Left);
+            self.delete();
+        }
     }
     fn delete(&mut self) {
         self.buffer.delete(self.text_location);
@@ -70,7 +78,7 @@ impl View {
             .map_or(0, Line::grapheme_count);
         let grapheme_delta = new_len.saturating_sub(old_len);
         if grapheme_delta > 0 {
-            self.move_right();
+            self.move_text_location(&Direction::Right);
         }
         self.needs_redraw = true;
     }
@@ -83,7 +91,6 @@ impl View {
         if height == 0 || width == 0 {
             return;
         }
-
         #[allow(clippy::integer_division)]
         let vertical_center = height / 3;
         let top = self.scroll_offset.row;
